refactor: removendo metodos e propriedades que não foram usados (validar e etc...)

+++ src/main/java/com/sonic/team/sonicteam/repository/CategoriaRepository.java
@@ -3,5 +3,5 @@ package com.sonic.team.sonicteam.repository;
 import com.sonic.team.sonicteam.model.CategoriaLivro;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoriaRepository extends JpaRepository<CategoriaLivro, Integer> {
+public interface CategoriaRepository extends JpaRepository<CategoriaLivro, Long> {
 }
